+++ CHANGES.rst
@@ -6,12 +6,13 @@ Changelog
 
 - Add budget.behavior
   [vpiret]
-
 - Corrected persistency bug in annotation when removing item.
   Don't set children budget at all on initial state.
   Removed all sub children on initial state.
   [sgeulette]
 - Managed move event to update children budget annotation
+  [sgeulette]
+- Used UID key in getVocabularyTermsForOrganization vocabulary
   [sgeulette]
 
 1.2 (2019-06-23)